Proper Trait definitions + some testing
Lets hope I got all that stupid lifetime and borrowing stuff right
this time.

Write it once to make it work, write it twice to make it good and
a third time to make it perfect. Well it works now that I did it
twice. Not sure what that says about my rust skills.